dashboard_service/requirements.txt
argcomplete==1.12.2
beautifulsoup4==4.9.3
boto3==1.16.59
botocore==1.19.59
bs4==0.0.1
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
durationpy==0.5
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
hjson==3.0.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
pip-tools==5.5.0
placebo==0.9.0
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-slugify==4.0.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.56.0
troposphere==2.6.3
urllib3==1.26.2
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
